# contracts/access-gate.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — Core Business Access Policy API
  version: 1.0.0
  description: |
    Hợp đồng API giữa Access Gate (Consumer) và Core Business (Provider).
    Pair 10: Access Gate gọi Core Business realtime để kiểm tra policy ra/vào trước khi mở cổng.

    Đàm phán bởi:
    - Consumer: Access Gate Service
    - Provider: Core Business Service
  contact:
    name: Pair 10 — Access Gate ↔ Core Business
    email: fit4110@example.edu.vn
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: https://api.campus.local
    description: Class network placeholder
...
security:
  - bearerAuth: []
paths:
...
  /access/check:
    post:
      operationId: checkAccessPolicy
      summary: Kiểm tra policy ra/vào
      description: |
        Access Gate gọi endpoint này mỗi khi có lượt quẹt thẻ.
        Core Business kiểm tra policy và trả ALLOW hoặc DENY.
        Response phải trả trong 500ms để tránh kẹt cổng.
      tags: [access]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessCheckRequest'
            examples:
              allowEntry:
                summary: Quẹt thẻ vào cổng chính
                value:
                  cardId: RFID-2026-001
                  gateId: GATE-01
                  direction: IN
                  timestamp: '2026-05-10T08:00:00Z'
                  idempotencyKey: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babe
              exitCheck:
                summary: Quẹt thẻ ra cổng phụ
                value:
                  cardId: RFID-2026-042
                  gateId: GATE-03
                  direction: OUT
                  timestamp: '2026-05-10T17:30:00Z'
                  idempotencyKey: null
      responses:
        '200':
          description: Quyết định ra/vào
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckResponse'
              examples:
                allowed:
                  summary: Cho phép vào
                  value:
                    decisionId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babf
                    decision: ALLOW
                    reasonCode: null
                    policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                    expiresAt: '2026-05-10T08:05:00Z'
                    checkedAt: '2026-05-10T08:00:00Z'
                denied:
                  summary: Từ chối vào
                  value:
                    decisionId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac1
                    decision: DENY
                    reasonCode: CARD_EXPIRED
                    policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                    expiresAt: null
                    checkedAt: '2026-05-10T08:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/access:
    get:
      operationId: listAccessPolicies
      summary: Lấy danh sách policy ra/vào
      description: Access Gate lấy danh sách policy với cursor-based pagination để đồng bộ cache hoặc audit.
      tags: [policies]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          required: false
          description: Lọc theo trạng thái policy.
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, ARCHIVED]
      responses:
        '200':
          description: Danh sách policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyPage'
              examples:
                firstPage:
                  value:
                    items:
                      - policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                        name: Giờ hành chính - Khu A
                        description: Policy cho khu A trong giờ hành chính
                        status: ACTIVE
                        rules:
                          - ruleType: TIME_BASED
                            ruleId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac2
                            startTime: '07:00'
                            endTime: '22:00'
                            allowedDays: [MON, TUE, WED, THU, FRI]
                        createdAt: '2026-01-01T00:00:00Z'
                        updatedAt: '2026-05-01T00:00:00Z'
                    nextCursor: null
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/access/{policyId}:
    get:
      operationId: getAccessPolicyById
      summary: Lấy chi tiết một policy
      description: Access Gate lấy chi tiết policy theo UUID để hiển thị thông tin hoặc debug.
      tags: [policies]
      parameters:
        - $ref: '#/components/parameters/PolicyId'
      responses:
        '200':
          description: Chi tiết policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessPolicy'
              examples:
                activePolicy:
                  value:
                    policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                    name: Giờ hành chính - Khu A
                    description: Policy cho khu A trong giờ hành chính
                    status: ACTIVE
                    rules:
                      - ruleType: TIME_BASED
                        ruleId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac2
                        startTime: '07:00'
                        endTime: '22:00'
                        allowedDays: [MON, TUE, WED, THU, FRI]
                      - ruleType: ROLE_BASED
                        ruleId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac3
                        allowedRoles: [STUDENT, LECTURER, STAFF]
                        restrictedZones: null
                    createdAt: '2026-01-01T00:00:00Z'
                    updatedAt: '2026-05-01T00:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /decisions:
    get:
      operationId: listDecisions
      summary: Lấy danh sách quyết định ra/vào
      description: Truy vấn lịch sử quyết định ra/vào với cursor-based pagination để audit.
      tags: [decisions]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách quyết định
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionPage'
              examples:
                firstPage:
                  value:
                    items:
                      - decisionId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babf
                        cardId: RFID-2026-001
                        gateId: GATE-01
                        direction: IN
                        decision: ALLOW
                        reasonCode: null
                        policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                        expiresAt: null
                        checkedAt: '2026-05-10T08:00:00Z'
                    nextCursor: null
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /decisions/{decisionId}:
    get:
      operationId: getDecisionById
      summary: Lấy chi tiết một quyết định
      description: Lấy chi tiết quyết định ra/vào theo UUID để audit hoặc tra cứu sự cố.
      tags: [decisions]
      parameters:
        - $ref: '#/components/parameters/DecisionId'
      responses:
        '200':
          description: Chi tiết quyết định
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessDecision'
              examples:
                allowDecision:
                  value:
                    decisionId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babf
                    cardId: RFID-2026-001
                    gateId: GATE-01
                    direction: IN
                    decision: ALLOW
                    reasonCode: null
                    policyId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
                    expiresAt: null
                    checkedAt: '2026-05-10T08:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PolicyId:
      name: policyId
      in: path
      required: true
      description: UUID của policy.
      schema:
        type: string
        format: uuid
    DecisionId:
      name: decisionId
      in: path
      required: true
      description: UUID của quyết định.
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor dùng để lấy trang tiếp theo.
      schema:
        type: [string, 'null']
        minLength: 1
        maxLength: 200
    Limit:
      name: limit
      in: query
      required: false
      description: Số item tối đa mỗi trang.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    AccessCheckRequest:
      type: object
      additionalProperties: false
      required: [cardId, gateId, direction, timestamp]
      properties:
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
          description: Mã thẻ RFID cần kiểm tra.
          example: RFID-2026-001
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã cổng yêu cầu kiểm tra.
          example: GATE-01
        direction:
          type: string
          enum: [IN, OUT]
          description: Hướng ra/vào.
          example: IN
        timestamp:
          type: string
          format: date-time
          description: Thời điểm quẹt thẻ.
          example: '2026-05-10T08:00:00Z'
        idempotencyKey:
          type: [string, 'null']
          format: uuid
          description: Key chống trùng request, null nếu không cần.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babe
    AccessCheckResponse:
      type: object
      additionalProperties: false
      required: [decisionId, decision, policyId, checkedAt]
      properties:
        decisionId:
          type: string
          format: uuid
          description: UUID của quyết định.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babf
        decision:
          type: string
          enum: [ALLOW, DENY]
          description: Quyết định cho phép hay từ chối.
          example: ALLOW
        reasonCode:
          type: [string, 'null']
          minLength: 2
          maxLength: 80
          description: Mã lý do từ chối, null nếu ALLOW.
          example: null
        policyId:
          type: string
          format: uuid
          description: Policy áp dụng cho quyết định.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: Thời điểm hết hạn quyết định, null nếu không giới hạn.
          example: '2026-05-10T08:05:00Z'
        checkedAt:
          type: string
          format: date-time
          description: Thời điểm kiểm tra policy.
          example: '2026-05-10T08:00:00Z'
    AccessPolicy:
      type: object
      additionalProperties: false
      required: [policyId, name, status, rules, createdAt, updatedAt]
      properties:
        policyId:
          type: string
          format: uuid
          description: UUID policy.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
        name:
          type: string
          minLength: 3
          maxLength: 120
          description: Tên policy.
          example: Giờ hành chính - Khu A
        description:
          type: [string, 'null']
          maxLength: 500
          description: Mô tả policy, null nếu không có.
          example: Policy cho khu A trong giờ hành chính
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
          description: Trạng thái policy.
          example: ACTIVE
        rules:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/PolicyRule'
        createdAt:
          type: string
          format: date-time
          description: Thời điểm tạo policy.
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Thời điểm cập nhật gần nhất.
          example: '2026-05-01T00:00:00Z'
    PolicyRule:
      oneOf:
        - $ref: '#/components/schemas/TimeBasedRule'
        - $ref: '#/components/schemas/RoleBasedRule'
      discriminator:
        propertyName: ruleType
        mapping:
          TIME_BASED: '#/components/schemas/TimeBasedRule'
          ROLE_BASED: '#/components/schemas/RoleBasedRule'
    TimeBasedRule:
      type: object
      additionalProperties: false
      required: [ruleType, ruleId, startTime, endTime, allowedDays]
      properties:
        ruleType:
          type: string
          const: TIME_BASED
          description: Loại rule.
          example: TIME_BASED
        ruleId:
          type: string
          format: uuid
          description: UUID của rule.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac2
        startTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: Giờ bắt đầu cho phép (HH:mm).
          example: '07:00'
        endTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: Giờ kết thúc cho phép (HH:mm).
          example: '22:00'
        allowedDays:
          type: array
          minItems: 1
          maxItems: 7
          items:
            type: string
            enum: [MON, TUE, WED, THU, FRI, SAT, SUN]
          description: Các ngày trong tuần cho phép.
          example: [MON, TUE, WED, THU, FRI]
    RoleBasedRule:
      type: object
      additionalProperties: false
      required: [ruleType, ruleId, allowedRoles]
      properties:
        ruleType:
          type: string
          const: ROLE_BASED
          description: Loại rule.
          example: ROLE_BASED
        ruleId:
          type: string
          format: uuid
          description: UUID của rule.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac3
        allowedRoles:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            minLength: 2
            maxLength: 50
          description: Danh sách role được phép.
          example: [STUDENT, LECTURER, STAFF]
        restrictedZones:
          type: [array, 'null']
          minItems: 0
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 50
          description: Vùng bị hạn chế, null nếu không có.
          example: null
    AccessDecision:
      type: object
      additionalProperties: false
      required: [decisionId, cardId, gateId, direction, decision, policyId, checkedAt]
      properties:
        decisionId:
          type: string
          format: uuid
          description: UUID quyết định.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babf
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
          description: Mã thẻ RFID.
          example: RFID-2026-001
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã cổng.
          example: GATE-01
        direction:
          type: string
          enum: [IN, OUT]
          description: Hướng ra/vào.
          example: IN
        decision:
          type: string
          enum: [ALLOW, DENY]
          description: Quyết định.
          example: ALLOW
        reasonCode:
          type: [string, 'null']
          minLength: 2
          maxLength: 80
          description: Mã lý do từ chối, null nếu ALLOW.
          example: null
        policyId:
          type: string
          format: uuid
          description: Policy đã áp dụng.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2bac0
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: Hết hạn quyết định, null nếu không giới hạn.
          example: null
        checkedAt:
          type: string
          format: date-time
          description: Thời điểm kiểm tra.
          example: '2026-05-10T08:00:00Z'
    PolicyPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          minItems: 0
          maxItems: 100
          items:
            $ref: '#/components/schemas/AccessPolicy'
        nextCursor:
          type: [string, 'null']
          description: Cursor trang tiếp theo, null nếu hết.
          example: null
        hasMore:
          type: boolean
          description: Còn dữ liệu hay không.
          example: false
    DecisionPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          minItems: 0
          maxItems: 100
          items:
            $ref: '#/components/schemas/AccessDecision'
        nextCursor:
          type: [string, 'null']
          description: Cursor trang tiếp theo, null nếu hết.
          example: null
        hasMore:
          type: boolean
          description: Còn dữ liệu hay không.
          example: false
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI định danh loại lỗi.
          example: https://campus.local/errors/validation
        title:
          type: string
          minLength: 3
          maxLength: 120
          description: Tiêu đề lỗi ngắn gọn.
          example: Dữ liệu không hợp lệ
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code.
          example: 400
        detail:
          type: [string, 'null']
          maxLength: 500
          description: Mô tả chi tiết lỗi.
          example: Field cardId không đúng định dạng
        instance:
          type: [string, 'null']
          format: uri
          description: URI của request gây lỗi.
          example: https://api.campus.local/access/check
        errors:
          type: array
          minItems: 0
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required: [field, code, message]
            properties:
              field:
                type: string
                minLength: 1
                maxLength: 80
                example: cardId
              code:
                type: string
                minLength: 2
                maxLength: 80
                example: PATTERN_MISMATCH
              message:
                type: string
                minLength: 3
                maxLength: 200
                example: cardId phải có dạng RFID-YYYY-NNN
  responses:
    BadRequest:
      description: Dữ liệu sai định dạng
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/validation
            title: Dữ liệu không hợp lệ
            status: 400
            detail: Payload không đúng JSON Schema
            instance: https://api.campus.local/access/check
            errors:
              - field: cardId
                code: PATTERN_MISMATCH
                message: cardId phải có dạng RFID-YYYY-NNN
    Unauthorized:
      description: Thiếu hoặc sai thông tin xác thực
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/unauthorized
            title: Chưa xác thực
            status: 401
            detail: Thiếu Bearer token
            instance: https://api.campus.local/access/check
            errors: []
    Forbidden:
      description: Không có quyền thực hiện thao tác
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Xung đột nghiệp vụ hoặc trùng idempotency key
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: Payload đúng định dạng nhưng không hợp lệ về nghiệp vụ
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Lỗi không mong muốn phía server
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'